Rzutowanie dalsza czesc

+++ cast_float.php
@@ -1,10 +1,10 @@
 <?php
 $var=5;
-$cast=(int) $var;
+$cast=(float) $var;
 var_dump($var,$cast);
 $var=5.5;
-$cast= (int)$var;
+$cast= (float)$var;
 var_dump($var,$cast);
 $var='4php';
-$cast=(int)$var;
+$cast=(float)$var;
 var_dump($var,$cast);
